Add settile() as an internal command
This works pretty well. Have to set `dwgfx.foregrounddrawn` to false in
order to force the game to redraw the room, though.

The problem is with spikes, one-way tiles, and 1x1 disappearing
platforms. You see, those aren't actually tiles - they're actually
blocks, and the tiles are just visuals. Well, the 1x1 disappearing
platforms' visuals actually come from the platform itself and not from a
tile, but still, there's nothing behind it in the tilemap.

I might add some kludge later to fix spikes, one-ways, and 1x1
disappearing platforms. I can create a block for these just fine, but
the annoying part is having to remove one if you're switching out, say,
a spike, to be a non-spiky tile...