Add customifnotflag() internal, ifnotflag() simplified
This command specifically checks if a flag is *off*, as opposed to it
being on, so you no longer have to do some kludge like

    ifflag(1,script)
    iftrinkets(1,flag_not_on)

All the usual properties of ifflag() apply. ifnotflag() is basically
just an inverted ifflag().

In internal scripting, ifnotflag() will load a main game script, but you
can use customifnotflag() to load a custom level script. In simplified
scripting, ifnotflag() will load a custom level script (like ifflag()
does, it just converts to customifnotflag() in the
simplified-to-internal parser). There are no bounds checks on what flags
you can check, so you can check out-of-bounds flags (and be able to
segfault the game by doing so), just like with ifflag(). Since ifflag()
checks if the integer is 1, I made sure that ifnotflag() checks if the
integer is *not* 1, and not if the integer is 0.

This was suggested by Shiny.