Make followposition work on player
changeai(player,followposition,<x>) now works.

It's the equivalent of a walk() - it causes a delay for however long it
takes Viridian to get over to that x-position, and then holds left or
right accordingly for you. Don't worry, if they get stopped along the
way, then the game doesn't softlock or anything, it'll run out
eventually.