Add customquicksave()
I've made it so you can quicksave from a script in a level! Useful for
having an auto-save function, of course.

The command will do nothing if you're playtesting in the editor, and
this is apparently when map.custommode is true but map.custommodeforreal
is false. So it looks like map.custommode corresponds with being in the
editor and map.custommodeforreal corresponds with being in-game.