Fix spikes, one-ways, and 1x1 quicksand with settile()
Whoa, that was a LOT of kludge in order to make them work!

Anyways, settile() now deletes the collision of spikes, one-way tiles,
and 1x1 disappearing platforms if there was such a tile on the tile you
targeted, and will now create the collision of spikes, one-way tiles,
and 1x1 disappearing platforms if you placed such a tile.