Fix enemies in Warp Zone gray tilecol not being gray
This is some kludge specifically for Warp Zone (tileset 3) gray (tilecol
6). The reason why the enemies are green is that since it's a tilecol
not in the range of 0..5, it starts using the color of the NEXT
tileset's tilecols, which just happen to be Ship green, and so the
enemies are green.

I've added a variable that denotes when to run the gray enemies kludge,
and it only happens if the tileset is 3 and the tilecol is 6. So Warp
Zone gray enemies now render as gray in the editor and in-game.