Fix formatting and wording in changelogs
There was one line that was too long to fit in the in-game changelog, so
I changed it. I made every line use the past-tense of words, i.e.
"Added" instead of "Add". I capitalized "Ved", and I clarified that the
sixth argument to `createentity` uses raw colors, and not
`createcrewman` colors.