Add createroomtext(x,y) -> next line
You can now create roomtext on-the-fly! Simply do something like this:

    createroomtext(x,y)
    Roomtext here

The X and Y are in tiles, not pixels.

This creates a single line of roomtext. To create multiple lines, use
multiple roomtexts.

destroy(roomtext) has been updated to set the number of roomtexts to 0,
instead of only turning roomtext off.