Add nobars as a simplified command
This means the game will no longer automatically insert cutscene bars
when it sees a `say` or `reply` in a script.

You will have to add the cutscene bars yourself if you are using this
command, but that should be okay since internal commands work in
simplified anyway.

This is mainly for one-time scripts where you want to keep the actual
script contents in the same script as the flag-checking, but you don't
want cutscene bars.