Add finalstretchon and finalstretchoff
It turns on and off the palette swap in the final stretch portion of the
Final Level.

I actually had to make the game animate the tiles if `map.finalstretch`
is on, because it only does it otherwise if `map.finalmode` is on, but I
didn't want to turn it on.