Add reloadscriptboxes
It reloads all the script boxes in the room that were touched by the
player and removed.

How this works is I added a new vector to store script boxes when they
get removed, and resurrect them when reloadscriptboxes is called. This
vector gets cleared on every gotoroom.

This works with destroy(scriptboxes), too - so you can do
destroy(scriptboxes) in a room, then do reloadscriptboxes to get them
back.