Fix rendering of quicksand/conveyors/platforms in finalstretch
Quicksand, conveyors, and platforms now use all the Warp Zone tiles from
entcolours.png.

Whoa, lots of kludge for this one! I especially spent a lot of time
fixing the rendering of the disappearing platforms, because there was a
problem if the room color changed while a disappearing platform
disappeared, and the most time was spent fixing the case when the room
color changed during disappearing platforms' respawning animation!