Fix broken fadein code (even w/ no 2nd arg) in play()
The code was targeting the first argument for the fade-in time, even
though the fade-in time is SUPPOSED to be the second argument. So, if
you had something like play(4), it would play track 4, but ALSO attempt
to start a fade-in of 4 milliseconds.

Now, there's a separate issue where it seems like fade-in lengths of 22
milliseconds or less result in not playing the song at all. I'm not sure
the cause of that one, but my get-it-done solution is to just have a
lower bound on the amount of milliseconds you can give to play() that is
23 milliseconds, which I think is reasonable and wouldn't get noticed by
anyone - so if we can't fix the root cause of this, we at least already
have a solution for it.