Add a labels and goto system
VVVVVV: CE scripting just got a whole lot powerful.

Here's the rundown: you define a label by typing $somethinglikethis$ on
a line. You cannot use spaces, parentheses, or commas. Using capitals is
not recommended. If you really want, you CAN embed a dollar sign inside
the label, but I don't know why you would want to.

Then, to jump to a label, all you have to do is use a command that takes
in a script as an input. iftrinkets(0,$thisismylabel$) means to jump to
the label thisismylabel in the current script, and
iftrinkets(0,anotherscript$anotherlabel$) means to jump to the label
anotherlabel in the script anotherscript.

If you attempt to jump to an invalid label in another script, nothing
will happen and the script will stop. If you attempt to jump to an
invalid label in the current script, it just goes to the top of the
script. So I guess you can use iftrinkets(0,$$) as a cheap trick to
restart the current script.

So here's an example:

    script_one:
    say
    This is text box #1
    iftrinkets(0,$skip_textbox$)
    say
    We skip over this text box
    $skip_textbox$
    say
    This is text box #2

If you run script_one, "This is text box #1" and "This is text box #2"
will appear, but NOT "We skip over this text box".

Using the same script, here's an example of jumping to another script's
label:

    script_two:
    iftrinkets(0,script_one$skip_textbox$)

If you run script_two, "This is text box #2" will appear, but not any of
the other text boxes.