Add ifvce()
This is so levels can formally detect if it is being run under VVVVVV:
Community Edition. If it is not, the game will not recognize the command
ifvce(), and will just keep executing the script. So if it jumps to the
script, the level knows it's running under VVVVVV-CE, and if it doesn't,
then it knows it's not.