Fix 2-frame-delay when executing a room's "init script"
Here, init script is defined as a script that creates something that, in
an ideal world, would naturally be placed as part of the room.

The reason for the 2-frame-delay in the first place is because when the
game does a gotoroom to another room, it has already checked whether or
not the player is standing inside a script box, which adds the first
frame of delay. Then when it does finally get around to checking
collision with script boxes, it actually sets the gamestate to something
between 300..336 depending on script box placement order, but it's
ALREADY processed the current gamestate for that frame, so it takes
ANOTHER frame before it finally actually executes the script. Thus, a
2-frame delay.

To fix this, I added some kludge such that the game will check if the
player is inside a script box right after it does a gotoroom, and bypass
the whole gamestate system and just execute the script immediately
(otherwise there would still be 1 frame of delay). However, this can
result in doing too much gotoroom, especially if you intend on entities
glitching out rapidly instead of having them be frozen. So, it'll only
do it for "natural" gotorooms, e.g. gotorooms that are NOT issued by a
script command.

Note that, since it DOES bypass the gamestate system, this kludge will
execute script boxes placed above the 37-usable-script-boxes limit even
though it's not supposed to do so. I don't particularly care about
fixing it, we haven't even run into the 37-usable-script-boxes limit in
practice.

This is probably the BIGGEST breaking of compatibility with levels, but
only levels where it matters that e.g. an enemy or platform is spawned 2
frames earlier or later. Such a frame-perfect level is probably
extremely unreasonably hard anyway.