Add support for loading altstates from level file
Altstates are stored in a new <altstates> tag, just after <contents>.
Each altstate looks like the following: <altstate x="" y=""
state="">...</altstate>, where X and Y are the coordinates of the room,
and state is the given alt state, and the thing in between is a bunch of
comma-separated concatenated-rows tiles.

Edentities in an alt state will have a 'state' attribute, and all
edentities that are on alt state 0 (no alt state) will NOT have a
'state' attribute. So an edentity in an alt state may look like the
following:

            <edentity x="23" y="19" t="18" p1="0" p2="0" p3="0" p4="0" p5="320" p6="240" state="1">a
            </edentity>

This commit adds the functionality of loading alt states into the game's
memory from a level file, as well as support for actually having alt
states in action once it is loaded from the level file.

To do:
 * Save levels with alt states
 * Add editor interface for alt states