Add position(centerx,<line>) and position(centery,<line>)
So now you can position the text box AROUND a given x= or y= line,
instead of being restricted to centering around the center of the
screen.

This wasn't that hard to do because the game ALREADY uses the same
thing, except for horizontally centering in the middle of the screen it
uses the line x=160, and for vertically centering in the middle of the
screen it uses the line y=120.

This feature was suggested by Shiny.