Revert "Made editorclass::absat() not use a pointer interface"
Nevermind, I forgot about Chesterton's Fence again. Instead of finding a
weird thing, asking why it's there, and UNDERSTANDING it before removing
it, I just went straight to "remove it".

Basically, the pointer thing is there because we DON'T want to make a
local copy of the X and Y.

This reverts commit 925d6de8f040be3c8c5308a14fff01073817dc31.