Fix altstates reset code when loading new level
Apparently, you can't call a constructor from a non-constructor function
and expect the constructor to execute and reset the current object's
attributes. So, I guess that's why every time I see a call to the
object's reset function inside a constructor, it's a call to reset()
inside the constructor and not a call to the constructor inside reset().
The more you know.