Remove check for nonzero altstate at all when creating entities
Ok, I'm not sure why I had to check for a nonzero altstate in the first
place when creating entities.

Anyway, this fixes a bug where entities from ALT states persist in the
main state, instead of entities in the main state persisting in alt
states. It's just a simple "if this entity's state matches the current
alt state" check, I shouldn't have bothered with anything more
complicated.