Add reloadcustomactivityzones()
This reloads all non-terminal activity zones in the room.

The code for reloadscriptboxes() has been shuffled around a bit to
accomodate the fact that removed script boxes are now no longer the only
ones in obj.resurrectblocks, and obj.cleanupresurrectblocks() has been
added as a function.

This command was suggested by Pixelator.