Add a failsafe to obj.removeblock() to not remove already-removed blocks
A bit of a downside with the "`active` system" (I like to call it) that
VVVVVV uses, removed objects aren't actually removed, and you need to
make sure that you check the `active` attribute of an object before
doing anything with it. Since I've added some code to resurrect removed
blocks, we'll need to start checking the `active` attribute even when
removing a block, so there's no possibility of duplicating removed
blocks in `obj.resurrectblocks`.