Add textboxtimer(n)
This command fades out the text box after n frames. It calls the same
function the main game calls for the "Game Saved" text box.

It is recommended to use this with a backgroundtext text box. This
command HAS to be run after a speak, speak_active, speak_fast, or
speak_active_fast, or else it won't work.