Fix Lab color mismatch (getlevelcol matching wrongly)
FIQ made editorclass::getlevelcol() check for the wrong tileset in
cec133456f9a6987c8d7ee77fd4674bbef0802c5 in the first case. He mixed up
the room property's tileset and the actual tilesheet in use.

Basically, the tilesheet numbers go like 0, 1, 2. These correspond with
tiles.png, tiles2.png, and tiles3.png. The room property tileset numbers
go from 0, 1, 2, 3, 4, and the newly-added 5. These correspond with
Space Station, Outside, Lab, Warp Zone, Ship, and the newly-added Tower
Hallways.

In cec133456f9a6987c8d7ee77fd4674bbef0802c5 he made it so the Space
Station and the Tower Hallways use the same case to return the enemy
color. Except, no, wait, he made it so the Space Station and the *Lab*
use the same case instead. This resulted in a mismatch between the color
of the tileset and the color of the enemies in the Lab.

He mixed the two up because he either didn't pay attention or was too
careless when adding the check to getlevelcol(). getlevelcol() checks
for the tileset room property, which goes like 0, 1, 2, 3, 4, 5, and not
the tilesheet number, which goes like 0, 1, 2. However, he mistakenly
made it check for tileset room property 2 instead of tilesheet number 2
(which would correspond with tileset room property 5 instead of 2), so
that commit made the Lab color mismatch. I'm certain of it, I checked
with git bisect and everything.

Anyway, this commit fixes the color mismatch. The enemy colors in the
Tower Hallways are still mismatched, though.