Fix printf and friends on Windows
I don't think we've run in to any of the related issues, but this should
preemptively fix them. MinGW uses Visual Studio 6.0 runtimes by default
(as these are the only versions that ship with older versions of
Windows), which are not standards-compliant. It has a workaround for
this that isn't activated by default. This isn't a problem on MSVC, as
newer MSVC runtimes are much more compliant.