Use strcpy_safe and strlcpy instead of strcpy
strcpy is undefined behavior if the source is longer than the
destination. Bruce Dawson (randomascii) created a C++ version of strcpy
that has correct behavior in all situations, but must be provided an
array (not a pointer). In the one situation where this is not the case,
libbsd's strlcpy (which strcpy_safe is simply a thin wrapper for) is
used instead.