Raise the map limit to 100x100
I had to find every place where the 20x20 limit was hardcoded and change
them to use ed.maxwidth and ed.maxheight.

The only thing I didn't really touch was roomdeathsfinal. That's kinda
for the Outside Dimension VVVVVV map, which for some reason is actually
20x20 despite being able to fit in a rectangle smaller than that.

And thankfully, the markers of teleporters and shiny trinkets don't use
the concatenated-rows system, and actually use separate X and Y
coordinates for storage. So there's that.