Don't save more room properties than necessary
The amount of items in each row of the concatenated-rows saved to the
level file is the same width as the width of the map, except if it is
less than 20, in which case it will be 20 because the original level
format always uses concatenated-rows of 20.

This makes <levelMetaData> behave more like <contents> in regards to the
width of each concatenated row.