Pad out extraneous room properties if map is 20x20 or less
This is so Ved won't throw an error if it gets a level with less than
400 room properties and the level is not exactly 20x20, because earlier
I changed it so that VVVVVV: CE only saves as many room properties as
necessary, but now I made it so it will pad it out if the map is 20x20
or less, i.e. not created for VCE only.