Add support for sub-tile positioning of edentities
Edentities now have 'subx' and 'suby' attributes, which are intended to
go from 0..7 (but it's actually unbounded, meaning you can do some crazy
stuff like having an edentity outside of the screen but still in the
same room), and is their precise pixel position. It's simply added to
their coordinates when the time comes to convert its tile position into
an actual pixel position.

There is no interface in the editor for this yet, not sure what a good
UI would be.