Fix music in custom level time trials not playing
The problem was that in 031e0675388df502f35cd498ee88480790803c84, Ally
decided to set music.currentsong when starting the time trial, but this
causes an issue. The variable is intended to keep track of the
currently-playing song, so that if the game is requested to play the
song again, it's not going to abruptly re-start the song in the middle,
which sounds a bit jarring.

However, since Ally set the current song to the currently-playing song,
the game will now think that it's ALREADY playing the time trial music,
even though that isn't the case. Thus, it will receive the music.play(),
decide that it's already playing the song, and will not actually play
the song.

The solution here is to set the current song to -1, i.e. not playing a
song, so that the game will properly start playing the time trial music.