Replace createactivityzone() with vanilla code
This fixes a bug where the game crashed if it encountered the vanilla
glitch to spawn a custom activity zone. Actually, the bug is the game
crashing if ANY invalid crewmate name was given to createactivityzone(),
regardless of whether that invalid crewmate name was a part of the
process to invoke said glitch or not.

In b0064c2eaf533af3335fb6a3558138253a3a7764, Leo decided to make
createactivityzone() use a version of obj.getcrewman() that took in a
string. This on its own didn't break anything. It was only in
fefbf4626089c337542e6678d39073a67a9369af that Leo added named crewmates
using that existing string-taking version of obj.getcrewman() where it
started to break. It was attempting to find an invalid crewmate name,
and it was returning uninitialized memory, which aborted the game when
the glitch was attempted to be invoked.

So now I'm putting back the vanilla code. This also has the benefit of
setting the variable `i`, which is essentially a class-wide variable
instead of being a temporary one. I've also put up a giant warning sign
telling people to NOT get any bright ideas about messing with
createactivityzone(). Just do customactivityzone() instead.