Fix warp token destinations not showing if not in current room or state
The cause of this bug was FIQ's commit
2fc237c89c0051ccb7fc741a5b6a212d01dd1f3f.

In that commit, FIQ decided to silently and subtly reduce the
indentation level of the check that makes sure entities are in the
current room and state, by first negating it then using a `continue`, so
that the ACTUAL contents of the if-block could be indented by one less
level.

Unfortunately, this cleverness also broke the rendering of warp token
destinations. That is because before, warp token destinations were NOT
inside the block that checks if the entity is in the current room, and
were in their own separate block - but by doing this clever
negate-the-conditional `continue` stuff, the warp token destination
block gets skipped over if its entrance is not in the same room and
altstate as well.

So I'm fixing this, as well as putting up yet another warning sign to
not break things again by being clever.