Add support sub-tile dimensions of script boxes and activity zones
Now you can have pixel dimensions that are not a whole multiple of 8 for
script boxes and activity zones. Their p3 attributes will be the
sub-tile width, and p4 will be the sub-tile height. They're supposed to
go from 0 to 7, but they can be whatever you want - after all, it only
gets added to their multiple-of-8 dimension.

Just like non-multiple-of-8 enemy and platform bounds, there isn't
support for this in the editor UI yet.