Add support for one-time script boxes
Script boxes can now have a `onetime` attribute. It is either 0 or
nonzero in the XML, true or false when it is loaded as an `edentity`,
and also true or false when it is a block inside `obj.blocks`. It will
only carry this attribute in the XML if it is true.

This commit will read and write the `onetime` attributes of edentities
to the level XML.

Here's how one-time script boxes work: when a player touches a script
box with a `onetime` set to true, the game will keep track of the script
name the script box was carrying, and store it in an array. Then, the
next time a `map.gotoroom()` happens, the game will NOT create a script
box if it is marked as one-time AND its script has already been ran.
Furthermore, when the game removes a one-time script box due to the
player touching it, it will look through the rest of the script boxes in
the room, and remove them if they have the same script AND if they are
ALSO one-time.

However, if a script box is NOT marked as one-time, then it will still
get created even if its script has already been ran. As well, when the
game removes a script box due to the player touching it, the game will
NOT remove non-one-time script boxes even if they have the same script
as the just-removed script box.

There is a new variable `obj.kludgeonetimescript` to signal not only
when a script box has been targeted for removal by a player (as opposed
to other means, like `destroy(scriptboxes)`) so that
`obj.removetrigger()` can then remove all the other one-time script
boxes in the room with the same script as the script box the player
stepped in, but it (`obj.kludgeonetimescript`) ALSO signals when a given
script box is a one-time script box in `obj.createblock()`.

I've also collapsed the cases to start a script from a script box in
Terry's State Machine (also known as `game.updatestate()`), cases
300..336, so they all use the same code and I don't have to duplicate
the code each time I make a change to how script boxes are ran when
touched. I've also made sure to make the change I did to script box
running to the 2-frame-delay kludge fix I put in `map.gotoroom()`.

This commit does NOT add reading or writing already-ran one-time scripts
to or from the level quicksave file, nor does it add support for placing
one-time script boxes in the editor.