Fix loading variables from level quicksave file
Wow, it's a miracle that I didn't get the game to freeze from loading a
level quicksave file earlier due to my mistake.

Basically, I was looking at the next element in the save file overall,
as opposed to the next element in the <variables> array. Subtle
difference. So it would only get the first variable in the array, and
furthermore if there actually was another element overall in the save
file, the game would enter an infinite loop and freeze.