Copy-paste warp token code to towerlogic
Copy-pasted, verbatim. Now warp tokens will warp you instead of
endlessly making the warping noise in towermode.

There is still a problem with them, though. If they are not at the very
start of the tower, they are most certainly guaranteed to teleport you
into the spikes at the edges of the screen or offscreen, killing you.
Fixing this would require doing something with the tower camera, which
I'm currently too lazy to do and would rather have done later.