Add vceversion attribute to level root element
The root element was <MapData version="2">, now it's
<MapData version="2" vceversion="1">. That means, the version attribute
is kept in place so VVVVVV doesn't try to read a version number that
doesn't exist, and programs reading level files will now have an
"official" way of determining whether or not a level is a CE level.
VVVVVV-CE could also use that version number to display a warning if
incompatible changes have been made to the format and it sees a level
with a version higher than it can support, but right now it does
nothing with it.

Feel free to change how this is indicated if for example you like
another name better, you like to use a different version numbering
scheme, if you like this to be a dedicated element, etc. As far as
I'm concerned the version number doesn't have to change before the
first release either, when no levels are being released yet.