Output on syslog for macOS
macOS silences all output from .apps. To work around this, we start
logger and redirect all output to it. This enables it to show up in
Console.app. This behavior can be disabled on macOS with --no-syslog,
and enabled on other platforms with --syslog. This disables normal
stdout, since the only way POSIX provides to mirror one fd into two
others is to use threads or the external tee command.