Fix vertically-moving entities' bounds being wrongly set
In 1a015ab4f60334933f9e892de6dda6f2675e3fa3, FIQ accidentally used
"warpx" instead of "warpy" in the check to increase the vertical bounds
of enemies and platforms if the room was warping. As a result, if the
room warped vertically or all-sides, the vertical bounds never got
increased, resulting in buggy vertical warping for those entities.