Account for camera when rendering warp token destinations
This fixes warp token destinations always rendering in the same
y-position onscreen and not moving with the camera. This also refactors
warp token destination rendering to use neat little variables `ep1` and
`ep2` so we don't have to type `edentity[i].p1*8` and `edentity[i].p2*8`
everywhere.

This fix exposes a problem that's always been here with warp token
destinations in towers - apparently, their position is anchored to the
top of the tower. So if you expand the tower upwards, the destination
moves along, too. But that's a separate issue to address, only now made
more apparent by this commit.