Only activate infiniflipkludge if infiniflip is on
This fixes a bug where if the player was both on the floor AND on the
ceiling at the same time, but infiniflip was NOT enabled, the player
would be able to switch between being flipped and unflipped at will,
even though the vanilla behavior is to only be able to end up unflipped
from being flipped and once you're unflipped you stay unflipped.

Some vanilla levels depend on this simultaneously-on-floor-and-ceiling
behavior, like my recent troll level, Traps R.J.!

But being simultaneously on the floor and ceiling doesn't mean only
being simultaneously on the floor and ceiling. It can also mean being
stuck in a wall, where the game sees you as being both on a floor and
ceiling, because, well technically you are.

Levels that take advantage of the fact that the game pushes you out when
stuck inside a wall, plus want to make sure you go a certain way, will
no longer have the guarantee that you will go a certain way if you can
switch to being flipped once you're unflipped when stuck inside a wall.
Dav999's level Dimension of Secrets takes advantage of the fact that you
cannot flip when you are unflipped inside a wall.