Draw roomname with text outline in gamerender
For some reason, roomnames simply weren't being drawn with text outline?
And in gamerender only, so roomnames were being drawn with their text
outline in towers?

Anyway, this meant that the roomname would be harder to read if you had
a translucent roomname background, which is not optimal, and is the
opposite of what is being done upstream, and is inconsistent with towers
which DOES draw the text outline.