Fix 2-frame-delay-fix code activating when resetting player
When the player respawns, map.resetplayer() gets called, which will call
map.gotoroom() if the player's checkpoint is not in the current room.

If map.gotoroom() is called here, the code to fix the 2-frame-delay
would activate BEFORE the player got re-positioned at their checkpoint.
If THAT happened, then the player could activate script boxes simply by
dying at their given position but in a different room, which is not what
we want.

I discovered this bug when playing through my latest level, Traps R.J.,
and finding that my out-of-bounds script boxes kept getting hit for some
inexplicable reason, before reasoning it through that it was due to this
happening.