Add death sequence check to 2-frame-delay-script fix
I think that another way to prevent bugs like the ones
5b38506645d8dad2cf89e945eedbcf933276f230 fixed is to use the same checks
the game does when it runs a script box. Namely, it first checks that
the player is currently not dying or in the middle of respawning.

It does this by having an if-else in gamelogic and towerlogic where the
if-branch checks that deathseq != -1, and in the else-branch,
obj.entitycollisioncheck() gets called, which at the very end of the
function checks if the player is standing inside a script box and
activates it if they are. So, since it only gets run in the else-branch
of the if-conditional deathseq != -1, I added the inverted check to the
2-frame-delay-script fix code, so it now checks that deathseq == -1.