Use Warp Zone textures for plats/quicksand/conveys in Tower Hallways
The plan is to add new rows of textures to entcolours.png for platforms,
quicksand, and conveyors inside towers and Tower Hallways. However, we
will re-use the Warp Zone textures as a fallback for (1) being a stopgap
measure in the short-term, and (2) when there is a custom entcolours.png
in place that doesn't have these extra rows.

I've added a new function ed.gettowerplattile() that returns the row in
entcolours.png given a Tower Hallways tilecol number, and updated
obj.createentity() and editorrender() to use this accordingly.