Respawn flip tokens upon death
When a flip token gets touched, it no longer gets deactivated. Instead,
its state will be set to 2, it will become invisible, and its 'onentity'
will be set to 0 (meaning it will no longer have collision). Then when
the player dies, its state will be set to 3, which means when the player
finishes dying, its state will be set back to 0, it will become visible
again, and its 'onentity' will be set to 1 (meaning it will now have
collision again).

This was suggested by Allie.