Draw teleporters in towers
I just copy-pasted the teleporter-drawing code from drawentities and put
it in drawtowerentities.

Unfortunately, it's drawn in front of the player, because in towers the
drawing order is reversed (drawn by entity order) from the drawing order
outside of towers (drawn by reverse entity order).

It sets your checkpoint and saves your game, but it's not much more
functional than that.