Fix moving platforms not colliding with conveyors in towers
The game's kludge to make moving platforms collide with conveyors is to
place tile 1s under them. Unfortunately, this wasn't happening in
towers.

I ran into three problems trying to remedy this:

(1) The tile-1-placing code was apparently inside the else-branch of an
    if-conditional checking that towermode was on.

    So I had to move it out of the if-else tree.

(2) The vector of arrays to modify is NOT `map.tower.contents`, but
    instead `map.tower.minitower`.

    This was the one that stumped me the most. `map.tower.contents`
    is used by the main game Tower, but `map.tower.minitower` is what is
    used for towers in custom levels, thanks to FIQ's implementation.

    Since the tile-1-placing code used map.settile(), I fixed in
    map.settile(). This also has the side effect of making the settile()
    internal command work properly in towers.

(3) You have to place down tile 12 instead of tile 1.

    Because in tiles3.png, tile 1 is not solid.