Fix SIGABRT from giving invalid crewmate name
named_crewmen.at(t) would throw if t is missing, and [t] returns
uninitialized memory, apparently.

To prevent this, we first check if it exists in the array before calling
named_crewmen.at(t). If it doesn't exist, we return 0, which is usually
the indice of the player entity, as the vanilla game does.