Invert and condense 2-frame-delay fix
This is a cleanup of the 2-frame-delay init script fix code.

I inverted the check that lets the 2-frame-delay kludge run. What I mean
is, instead of allowing it everywhere and making exceptions for when
it's not going to run, I'm instead taking the reverse approach, where I
*dis*allow it everywhere and only let it run in certain circumstances,
i.e. the screen transition code in Logic.cpp (in both gamelogic() and
towerlogic()).

This removes the need for the variables 'kludgeroominitscript' and
'gotoroomfromscript' on the 'game' object, so those have been removed.
(The 'kludgeroominitscript' in Logic.cpp is just a variable local to
gamelogic().) It makes the logic surrounding the 2-frame-delay fix much
more local, which increases its maintainability.

I've also abstracted it to a standalone function held in Map.cpp,
twoframedelayfix(). Note that it is standalone, so it is NOT an
attribute of the 'map' object.