Actually dynamically read and write <vceversion>
There's now a macro for VCE version, 'VCEVERSION', because I had to use
the actual version number more than once.

VCE will now save the version number read in the level file to the
internal variable 'ed.vceversion', which it might use to do things
differently.

The number version VCE will write to the level file is the version
number of VCE, and it will not maintain the version number in the level
file.

A version number of 0 usually means a non-existent <vceversion> tag,
which means it's a vanilla level.