Add check to 2-frame-delay fix that script box is custom script
The 2-frame-delay fix will only run if the script box that the player is
standing in is a custom script, i.e. its trigger number is 300 or
higher. Main game script boxes' trigger numbers are always less than
that.

It would be a problem if there were a main game script box, because I
have to subtract the trigger number by 300 and index game.customscript
to get the script the script box contains, and if the trigger number is
less than 300, it would result in negative array indexing, which would
most certainly throw a SIGABRT.