Fix setting vceversion when loading non-VCE level
Apparently, neither map.custommode nor map.custommodeforreal are set
until after the level is loaded, so we can't rely on those.

Instead, I'm just going to set the VCE version to the current one when
playtesting, and only when in-game and not playtesting will the VCE
version of the level be adhered to.

Also, I have to check if the 'vceversion' attribute exists or else it
seems like TinyXML just won't change 'ed.vceversion' to 0 or anything.