Add offsets to exposed internal variables
The concept is simple enough - the 'X' macros used to keep a list of
internal variables have a slight shortcoming where since the internal
variable being pointed to has to be an lvalue along with being an
rvalue, we can't do something like 'varname-50' because that's not a
valid lvalue.

My first thought was to do some pointer tricky, but I'm not sure that's
possible.

My second thought was to add another argument to the 'X' macro, which is
simply the offset of the variable in question.

Unfortunately the problems don't end there. Most of the variables don't
have an offset (i.e. they have an offset of 0), so they'd just be
lugging around an extra argument that is superfluous. However,
optional/default arguments and variadic arguments in general in macros
is very poorly supported. I'd have to either:
 (1) Add some monstrosity of a contraption of macros in order to get
     "clean" variadic-argument 'X' macros, or
 (2) Add another macro for variables with offsets, which requires yet
     another definition and un-define every single time SPECIALVARS is
     used, looking like just a bunch of duplicated code and making it a
     bit ugly

In the end, I decided that it was better to just add an extra 0 to every
variable, even if in an ideal world we would want that extra argument
gone for variables that don't have an offset.

To aid maintainability, I added a comment to the SPECIALVARS define that
makes it clear the third argument is offset/indexing, not like a default
value or something.