Add %room_x% and %room_y% as special scripting variables
These variables provide a direct interface to access the coordinates of
the current intended room, game.roomx and game.roomy.

Note that changing these won't update the room - these are the
coordinates of the current INTENDED room, not the actual current room.
You should use gotoroom() for that (it will help to use the relative
tilde syntax if you don't need anything more complicated than addition
or subtraction to a coordinate).

However, changing them still has an effect. If you change the
coordinates, the game will then use the new coordinates to calculate
which room to go into upon the next screen transition, instead of the
actual coordinates of the current room. So if you're in (0,0), then set
%room_x% to 1, then when you go to the bottom screen border, the game
will first think you're in (1,0), so it will move you to (1,1) instead
of (0,1).

Furthermore, if you change it and it's no longer the coordinates of the
actual room you're in, whenever you die the game will do a gotoroom() to
the current room. So there's a neat trick there.