Make flip token only play gravity line sound when touched
It used to be if you were eligible to flip off a floor or ceiling while
touching a flip token, it would play a flip or unflip sound. Which
sounds weird, mostly because a flip token doesn't make Viridian push off
of the floor or ceiling. But not only that, this could lead to the
flip/unflip sound being played twice, or even on the same frame at once
(which results in a loud flip/unflip sound which destroys your eardrums
like that high note in the song with that terrible ear-destroying note
hour-long note ( https://www.youtube.com/watch?v=Cu1gOAlvPx8 )).