Add parsebool() utility function
It's so whenever a boolean is used in a function, we don't have to keep
copy-pasting `words[1] == "true" || words[1] == "1"` everywhere.

Here are the valid truth values:
 - true
 - yes
 - y
 - on

Here are the valid false values:
 - false
 - no
 - n
 - off

If it doesn't match any of those, the argument gets converted to an int
through ss_toi(). And ss_toi() will just return 0 if it doesn't start
with a number or if it's empty. If the integer is nonzero, it's true,
otherwise it's false.